Add Three New Tests, Including One Temporary
First, "should be able to get available masks" tests that
MaskableAttribute has a masks method. The second, "should get available
masks" tests the actual result.

Lastly a temporary test for determining how to get the instance of
ActiveRecord to be apply the masks to the value of the masked attribute.
This and the method it tests, masked_object, are only temporary and will
be deleted (probably) after the test passes.