Complete set Method, Add get_value_of(mask) Method
Finish the set method to search for masks' values in the value being
assigned and subistute them back in to the value so that the masks' are
perserved.

Add get_value_of(mask) method to DRY up the process of getting the
value of masks from the @object. Refactor the masked method to use this
new method.