Accept New Options Hash Format
Options of hash can now specify a method key with value being a
symbol for a method to send to the masked object or a Proc to be
called with the masked object passed as an attribute. Formats are
still a hash. This can be used in lieu of an options array with
the method being the first element and the format options being the
second element.