Add Test for Non-Maskable Prefixes
Add a failing test to demonstrate handling attributes that have
non-maskable prefixes specified so that the attribute is left alone if
it matches the prefix. For example, if a non-maskable attribute prefix
of 'test' is specified, and the attribute has a value of
'testSomething', then the attribute will not be masked due to it having
'test' as its prefix.